feat: provide mock context for unit test
